tests: reorganize endtoend tests
